thread pool and corresponding benchmarks

+++ src/thread_pool/naive.rs
@@ -8,7 +8,7 @@ use crate::Result;
 pub struct NaiveThreadPool;
 
 impl ThreadPool for NaiveThreadPool {
-    fn new(_threads: u32) -> Result<Self> {
+    fn new(_threads: usize) -> Result<Self> {
         Ok(NaiveThreadPool)
     }
